requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.3
Flask==2.1.3
Flask-Alembic==2.0.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
pkgutil-resolve-name==1.3.10
pycparser==2.21
PyJWT==1.7.1
pyrsistent==0.18.1
pytz==2022.5
six==1.16.0
SQLAlchemy==1.4.42
typing-extensions==4.4.0
waitress==2.1.2
Werkzeug==2.1.1
zipp==3.9.0
